projectAssignment and initialAction swap
