+++ frontend/src/pages/projectAssignment.js
@@ -1,13 +1,5 @@
 import React from "react";
-import { makeStyles, withStyles, Typography, Box, Button, Grid } from "@material-ui/core";
-import Checkbox from "./components/checkbox";
-
-// const useStyles = makeStyles((theme) => ({
-//   root: {
-//     backgroundColor: theme.palette.background.paper,
-//     color: "#5b7f95"
-//   },
-// }));
+import { withStyles, Typography, Box, Grid, Button } from "@material-ui/core";
 
 const TextTypography = withStyles({
   root: {
@@ -15,7 +7,8 @@ const TextTypography = withStyles({
   }
 })(Typography);
 
-function projectAssignment({pages, setPages, activePage, setActivePage}) {
+function ProjectAssignment({pages, setPages, activePage, setActivePage}) {
+
   function goToIntroduction(){
     if (!pages.introduction.visited) {
       setPages(prevPages => {
@@ -26,16 +19,16 @@ function projectAssignment({pages, setPages, activePage, setActivePage}) {
     }
     setActivePage(prevPage => 'introduction')
   }
-  let goToResponseOne = ()=>{
-    if(pages.responseOne.completed){
-      if(!pages.responseOne.visited) {
+  let goToInitialAction = ()=>{
+    if(pages.initialAction.completed){
+      if(!pages.initialAction.visited) {
           setPages(prevPages => {
             let copy = {...prevPages};
-            copy.responseOne.visited = true;
+            copy.initialAction.visited = true;
             return copy;
           });
         }
-        setActivePage(prevPage => 'responseOne')
+        setActivePage(prevPage => 'initialAction')
     }
   }
 
@@ -44,16 +37,16 @@ function projectAssignment({pages, setPages, activePage, setActivePage}) {
       <Grid container direction="row" justify="center" alignItems="center">
         <Box mt = {5}>
           <TextTypography variant="h4" align="center" gutterBottom>
-          What should we not automate? 
+            the first response
           </TextTypography>
         </Box>
       </Grid>
       <Grid container direction="row" justify="space-between">
         <Grid item style={{ marginRight: "0rem", marginTop: "-3rem" }}>
-          <Button variant="contained" disableElevation onClick={goToIntroduction}>Back</Button>
+        <Button variant="contained" disableElevation onClick={goToIntroduction}>Back</Button>
         </Grid>
-        <Grid item style={{marginRight: "0rem", marginTop: "-3rem" }}>
-          <Button variant="contained" disableElevation color="primary" onClick={goToResponseOne} >Next</Button>
+        <Grid item style={{ marginRight: "0rem",  marginTop: "-3rem" }}>
+          <Button variant="contained" disableElevation color="primary" onClick={goToInitialAction}>Next</Button>
         </Grid>
       </Grid>
       <Grid container spacing={2}>
@@ -61,15 +54,12 @@ function projectAssignment({pages, setPages, activePage, setActivePage}) {
           <Box m="2rem">
           </Box>
           <TextTypography variant="body1" align="center">
-            Should there be some occupations or parts of human life where Artificial Intelligence and the use of robots are prohibited? 
+            Some Response.
           </TextTypography>
         </Grid>
-        <Grid item lg={12}>
-          <Checkbox nextPage = {goToResponseOne} pages={pages} nextPageName={"responseOne"}/>
-          </Grid>
       </Grid>
     </div>
   );
 }
 
-export default projectAssignment;
+export default ProjectAssignment;
